HW-1 BaseFragment api 요청 함수 추가, 탭별type 추가
