+++ developerkhm/AndroidStudy/app/src/main/java/com/example/androidstudy/ui/blog/BlogFragment.kt
@@ -2,23 +2,17 @@ package com.example.androidstudy.ui.blog
 
 import android.os.Bundle
 import android.text.TextUtils
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.bumptech.glide.Glide.init
 import com.example.androidstudy.R.layout
-import com.example.androidstudy.api.RetrofitBuilder
-import com.example.androidstudy.api.data.TotalModel
 import com.ironelder.androidarchitecture.view.AdapterBlog
 import kotlinx.android.synthetic.main.layout_search_view.*
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
-class BlogFragment : Fragment() {
+class BlogFragment : Fragment(){
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -66,18 +60,4 @@ class BlogFragment : Fragment() {
     }
 
 
-    private fun apiFetchData(str: String) {
-        val result = RetrofitBuilder.instance().requestSearchForNaver("blog", str)
-        result.enqueue(object : Callback<TotalModel> {
-            override fun onFailure(call: Call<TotalModel>, t: Throwable) {
-                Toast.makeText(context, t.message, Toast.LENGTH_SHORT)
-            }
-
-            override fun onResponse(call: Call<TotalModel>, response: Response<TotalModel>) {
-                var resultList = response.body()
-                Log.d("TEST1234", "RESUTL : ${resultList.toString()}")
-                (resultRecyclerView?.adapter as AdapterBlog).setItemList(resultList?.items)
-            }
-        })
-    }
 }
